refactor(install): split link phase
